fixes here and there. restoring ogosselet's original aixm_parser.py
