--- airspace/tmp.py
@@ -1,5 +1,0 @@
-from lxml import etree
-
-data = etree.parse('./tests/aixm_4.5_extract.xml')
-root = data.getroot()
-print(len(root.findall('Ase')))
